Rename the test from realtiup to realcluster
