+++ tests/integrations/realcluster/wait_tiup.sh
